requirements.txt
# Core dependencies
numpy>=1.26.4
Pillow>=11.0.0
opencv-python>=4.11.0.86
gradio>=5.29.1
openai>=1.54.5
pyyaml>=6.0.2
beautifulsoup4>=4.12.3
requests>=2.28.2
torch>=2.5.1
transformers>=4.47.1
diffusers>=0.32.2
tqdm>=4.65.2
python-dotenv>=1.0.1
pandas>=2.2.1
scikit-learn>=1.4.1
matplotlib>=3.8.3
seaborn>=0.13.2